Add a test to make sure the image shape is correct
I spent a lot of time troubling on the image loading reshaping the array
or transposing or not. I found that when I made the fake image being the
consecutive integer array, this would never be tested. Currently the
best way is to add a assertion to make sure the shape is right. Later I
should modify the fake image to ensure correct reshaping.