Modify state trace import for new model selection
I changed the imscroll selectK criteria to release the at least 20 frame
constraint to fit more rare and transient events.

But then in a trace with two state, the labels can be like (2, 4), then
the state analysis will be failing since it assumes consecutive state
labels. Fixed this by deflate the labels in advance.