Transform bi-exp likelihood function to log-domain
I noticed that there are some precision loss problem when doing the ML
estimation since scipy keeps telling me there are some log(0)
evaluation.

I found that the A parameter might contribute the most, since it can be
as close to 0 or 1, and (1-A) can have rounding error. So I rewrite the
log-likelihood function, transforming all the parameters to the log
domain to minimize the precision loss.