Make frame averaging into the ImageView widget
Override the updateImage() method to read a new averaged image from file
every time it is called, instead of pre-loading the whole stack from the
start. I need to patch up the 'z' axis to trick the widget to behave
like it is using a stack instead of a single frame.

The older average button and reset button were deleted.