Replace xlrd with openpyxl engine in read_excel
The xlrd package is not maintained anymore, which means in later python
version it may not be supported. There was a deprecation warning always
showing up when using pytest, so I think it is reasonable to replace it
now. However, the pandas read_excel goes to read_only mode when using
openpyxl, which calculates the wrong sheet size. So I patched the
read_excel function to avoid using read_only and placed it in utils.py.